Добавил Типы (Types)

+++ src/main/java/com/example/demo/persistence/model/VehicleTypeDto.java
@@ -28,7 +28,7 @@ public class VehicleTypeDto {
 	/**
 	 * Поля объекта, соответствующие полям таблицы в БД.
 	 */
-	int count;
+	long count;
 
 
 }
